% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_gen.R
\name{cg_indent}
\alias{cg_indent}
\title{Indent a character vector}
\usage{
cg_indent(x, indent_level = 0, indent_character = " ")
}
\arguments{
\item{x}{a character vector}

\item{indent_level}{the level to indent}

\item{indent_character}{the character to use for indenting}
}
\value{
an indented character vector
}
\description{
Indent a character vector
}
