% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_gen.R
\name{cg_surround}
\alias{cg_surround}
\title{Surround a string}
\usage{
cg_surround(
  x,
  surround = "()",
  style = c("single", "sandwich"),
  inner_indent = 0
)
}
\arguments{
\item{x}{the string to surround}

\item{surround}{enclosing characters. See details.}

\item{style}{the style of surrounding. See details.}

\item{inner_indent}{the indent of the inner part of the sandwich.
Only used when `style = 'sandwich'`.}
}
\value{
a surrounded string
}
\description{
Surround a string
}
\details{
The argument to `surround` should be 1 or 2 characters. If 1, then the
  returned string will be enclosed in that character. If 2, then
  the returned string will be enclosed by the first on the left and the second
  on the right.
See examples for the style argument.
}
\examples{
x <- month.abb[1:3]
cg_surround(x, '"')
cg_surround(x, style = 'sandwich')
cg_surround(x, '{}', 'sandwich', 2)
}
