% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_sets.R
\name{groupingsets}
\alias{groupingsets}
\alias{gs_bool_to_group_number}
\alias{grouping_number}
\alias{grouping_parser}
\alias{filter_grouping_set}
\title{Grouping sets}
\usage{
gs_bool_to_group_number(x)

grouping_number(all_cols, included_cols)

grouping_parser(all_cols)

filter_grouping_set(
  .x,
  all_grouping_cols,
  group,
  .group_number_col = grouping_number
)
}
\arguments{
\item{x}{a logical vector to convert into a grouping set number}

\item{all_cols}{all of the grouping columns}

\item{included_cols}{the group with the desired columns}

\item{.x}{a data.frame-like object}

\item{all_grouping_cols}{the columns included in the grouping sets}

\item{group}{the columns included in the desired group}

\item{.group_number_col}{the column containing the grouping number}
}
\value{
a single number indicating the grouping set corresponding to the
  bool

the grouping number

a function that will parse grouping columns

the parsed data frame
}
\description{
Functions for working with Grouping Sets, specifically as specified
in the [Presto documentation](https://prestodb.io/docs/current/sql/select.html#group-by-clause).
}
\details{
The grouping set number will correspond to the output of
 `grouping` in Presto.
}
\examples{
gs_bool_to_group_number(c(TRUE, FALSE, FALSE))
grouping_number(c("origin_state", "origin_zip", "destination_state"),
"origin_state")
}
