% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{guess_labeler}
\alias{guess_labeler}
\title{Guess which labeler to use for stat_annotate_}
\usage{
guess_labeler(x)
}
\arguments{
\item{x}{a vector to be labeled}
}
\value{

}
\description{
Given a vector, guess how to label it. This is used in e.g.
stat_annotate_min
}
