% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_cli.R
\name{new_paper}
\alias{new_paper}
\title{Create a paper entry}
\usage{
new_paper(x, extra_tags = NULL, directory = "notes", browse = TRUE)
}
\arguments{
\item{x}{the citation in Vancouver style}

\item{extra_tags}{extra tags}

\item{directory}{directory to create the notes}
}
\value{
The uri
}
\description{
Create a paper entry
}
