% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice-dice.R
\name{slice_top_categories}
\alias{slice_top_categories}
\title{Slice a dataframe to include only the top categories}
\usage{
slice_top_categories(.data, .n, ..., .wt = NULL)
}
\arguments{
\item{.data}{a dataframe}

\item{.n}{the number of categories to include}

\item{...}{columns which define categories}

\item{.wt}{an optional weight category}
}
\value{
A dataframe containing only the top categories
}
\description{
This slices a dataframe to only include the most frequent
}
\examples{
# give the top 1 most occuring value of `cyl`
slice_top_categories(mtcars, 1, cyl)

# give the top 1 most occuring value of `cyl` weighted by am
slice_top_categories(mtcars, 1, cyl, .wt = am)

# give the 2 most occuring values of (cyl, gear) pair
slice_top_categories(mtcars, 2, cyl, gear)
}
