% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme1}
\alias{theme1}
\title{ggplot themes}
\usage{
theme1(grid_type = c("both", "horizontal", "vertical", "none"))
}
\arguments{
\item{grid_type}{which grid lines should be displayed?}
}
\value{
a ggplot theme
}
\description{
Just ggplot2 themes that I like and think look good.
}
